requirement.txt
torch>=1.9.0
sentence-transformers>=2.2.2
numpy>=1.21.0
scikit-learn>=1.0.0
tqdm>=4.64.0
pyairtable
requests
python-dotenv